+++ game/sprites/cloud.py
@@ -10,7 +10,7 @@ class Cloud(sprite.Sprite):
         self.__game_env = game_env
         self.surf = image.load(random.choice(self.__game_env.static.clouds)).convert()    # loading cloud image
         self.surf.set_colorkey((255, 255, 255), self.__game_env.RLEACCEL)                           # setting the white color as the transperant area; RLEACCEL is used for better performance on non accelerated displays
-        self.__speed = 5
+        self.__speed = 4
         pos_x = random.randint(self.__game_env.static.screen_width + 10, self.__game_env.static.screen_width + 50)
         pos_y = random.randint(0, self.__game_env.static.screen_height - self.__game_env.vegetation_size[1] / 2)
         self.rect = self.surf.get_rect(center=(pos_x, pos_y))                                       # create rectange from the cloud screen