+++ game/sprites/sam.py
@@ -8,7 +8,7 @@ class Sam(sprite.Sprite):
         super(Sam, self).__init__()
         self.__game_env = game_env
         self.__angle = math.atan2(target[1] - source[1], target[0] - source[0])     # sam angle of fire in radian
-        self.__speed = 5 + self.__game_env.dynamic.game_level                       # default sam speed is 5 and increased each level
+        self.__speed = 5 + (1 if self.__game_env.dynamic.game_level%2 == 0 else 0)  # default sam speed is 5 and increased each level
 
         if flip:                                                                    # sam image rotational angle based on the fire position
             rotation_angle = 90 - self.__angle * (180 / 3.1415)