+++ CHANGELOG.md
@@ -1,7 +1,12 @@
 # Changelog
 
 ## [1.0.4] - Unreleased
-
+- added exit prompt in the game, this also acts as a game pause feature
+- changed size of sam missile
+- changed the probabity of a sam-launcher firing from 50% to 33%
+- changed the sam speed increase, now speed increases by 1 every 2 level instead of 1 level
+- changed the cloud and vegetation speed by increasing the vegetation's speed and reducing the cloud's speed
+- renamed ReplayText to ReplayMenuText
 
 ## [1.0.3] - 2020-05-10
 - added support for new api